% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{buscar_textos}
\alias{buscar_textos}
\title{Buscar textos}
\usage{
buscar_textos(tag)
}
\arguments{
\item{tag}{Tag buscada}
}
\value{
Uma tibble com as colunas titulo, data, texto e link
}
\description{
Buscar textos
}
\examples{
\dontrun{buscar_textos("ministerio-da-agricultura")}
\dontrun{buscar_textos("República Dominicana")}
}
