% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{listar_paginas}
\alias{listar_paginas}
\title{Listar paginas de busca}
\usage{
listar_paginas(tag)
}
\arguments{
\item{tag}{Tag buscada}
}
\value{
Um vetor com links das paginas da busca
}
\description{
Listar paginas de busca
}
